bcl: channel_to_vht_center_freq(): take eWifiBandwidth as argument
The channel_to_vht_center_freq function has a bandwidth argument. It is
currently taking an integer, which doesn't allow for 80+80 bandwidth.
When converting the 80+80 to integer, it is set as 160, which is not
really wrong, but it is not right for this function.

Most callers anyway already have the bandwidth as eWifiBandwidth, so
this actually simplifies the code.

While we're at it, also handle the 80+80 case, and report the value of
an invalid bandwidth.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>